[ UPDATE ] Removing PoC from LandingPage
