+++ src/components/index.ts
@@ -1,2 +1,8 @@
-export * from './theme/ToogleThemeButton'
-export * from './ui/spinner/AuthSpinner'
+// Base components
+export * from "./ui/avatar/Avatar";
+export * from "./ui/button/Button";
+export * from "./ui/input/Input";
+
+// Extra
+export * from "./theme/ToogleThemeButton";
+export * from "./ui/spinner/AuthSpinner";
