+++ src/pages/LandingPage.tsx
@@ -2,10 +2,8 @@ import { Link } from "react-router-dom";
 import { Waves1SVG, RocketSVG } from "../assets";
 import { DownArrowIcon } from "../assets/icons/DownArrowIcon";
 import { ToogleThemeButton } from "../components/theme/ToogleThemeButton";
-import { Avatar, Button } from "../components/ui";
-import { Input } from "../components/ui/input/Input";
+import { Avatar } from "../components/ui";
 import { User } from "../context/auth";
-import { useForm } from "../hooks/useForm";
 
 const testInitialUser2: User = {
   names: "Jorge",
@@ -16,21 +14,7 @@ const testInitialUser2: User = {
   token: "",
 };
 
-const formInitialState = {
-  email: "",
-};
-
-type FormState = {
-  email: string;
-};
-
 export const LandingPage = () => {
-  const { onInputChange, email } = useForm<FormState>(formInitialState);
-
-  const onSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    console.log(email);
-  };
   return (
     <main className="min-h-screen bg-light dark:bg-dark overflow-x-hidden transition-colors">
       <nav className="h-[10vh] flex items-center justify-around">
@@ -62,18 +46,7 @@ export const LandingPage = () => {
         />
       </div>
       <ToogleThemeButton fab={true} />
-      <div>
-        <form action="" onSubmit={onSubmit}>
-          <Button text="Hola" variant="secondary" className="mx-3 w-32" />
-          <Input
-            className="w-1/2"
-            name="email"
-            onChange={onInputChange}
-            value={email}
-            placeholder="Email"
-          />
-        </form>
-      </div>
+      <div></div>
     </main>
   );
 };
