+++ src/pages/LoginPage.tsx
@@ -1,7 +1,31 @@
-import React from 'react'
+import { Input, ToogleThemeButton } from "../components";
+import { useForm } from "../hooks/useForm";
+
+const formInitialState = {
+  email: "",
+  password: "",
+};
+
+type FormState = typeof formInitialState;
 
 export const LoginPage = () => {
-    return (
-        <div>LoginPage</div>
-    )
-}
+  const { onInputChange, email, password } =
+    useForm<FormState>(formInitialState);
+
+  const onSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    console.log(email, password);
+  };
+
+  return (
+    <form className="bg-light dark:bg-dark" onSubmit={onSubmit}>
+      <Input
+        placeholder="Email"
+        name="email"
+        value={email}
+        onChange={onInputChange}
+      />
+      <ToogleThemeButton fab={true} />
+    </form>
+  );
+};
