+++ src/pages/RegisterPage.tsx
@@ -1,0 +1,3 @@
+export const RegisterPage = () => {
+  return <div>RegisterPage</div>;
+};
